refactor(mooncake): own executable package resolution
